// src/app/home/models/product.model.ts
export interface Product {
  mainImage?: string;
  styleName?: string;
  releaseName?: string;
  status?: string;
  price?: number;
  salePrice?: string;
  currency?: string;
  brandImage?: string;
  productDate?: Date | number;
  slug?: string;
  hot?: boolean;
  shopName?: string;
  salePercentage?: number;
  sku?: string;
  thumbnail?: string;
  brandName?: string;
}